requirements.txt
# Core Dependencies
flask>=2.3.0
flask-socketio>=5.3.0
ollama>=0.1.7
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0

# Memory & AI
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# WebUI
flask-cors>=4.0.0
jinja2>=3.0.0
werkzeug>=2.0.0

# Phase 3 Integrations
discord.py>=2.3.0
websockets>=12.0
aiohttp>=3.8.0
piper-tts>=1.2.0  # ← ADD THIS LINE

# Emotional Analysis
textblob>=0.17.1

# Utilities
colorama>=0.4.6
click>=8.1.0
jsonschema>=4.17.0